experimental/RF/requirements.txt
absl-py==0.7.1
astor==0.7.1
backcall==0.1.0
certifi==2019.3.9
chainer==5.3.0
chardet==3.0.4
cloudpickle==0.8.1
cntk-gpu==2.6
contextlib2==0.5.5
cupy==5.3.0
cycler==0.10.0
Cython==0.29.6
dask==1.1.5
decorator==4.4.0
dm-sonnet==1.29
enum34==1.1.6
fastrlock==0.4
filelock==3.0.10
future==0.17.1
gast==0.2.2
graphviz==0.10.1
grpcio==1.19.0
h5py==2.9.0
idna==2.8
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
joblib==0.15.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Lasagne==0.2.dev1
leveldb==0.194
Mako==1.0.8
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mxnet-cu90==1.4.0.post0
networkx==2.2
nose==1.3.7
numpy==1.16.2
onnx==1.4.1
pandas==0.24.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
pycurl==7.43.0
Pygments==2.3.1
pygobject==3.20.0
pygpu==0.7.6
pyparsing==2.3.1
python-apt==1.1.0b1+ubuntu0.16.4.2
python-dateutil==2.8.0
python-distutils-extra==2.39
python-dotenv==0.13.0
python-gflags==3.1.2
pytz==2018.9
PyWavelets==1.0.2
PyYAML==5.1
requests==2.21.0
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
semantic-version==2.6.0
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
tensorflow-probability==0.6.0
termcolor==1.1.0
Theano==1.0.4+unknown
thundersvm==0.3.12
toolz==0.9.0
torch-nightly==1.0.0.dev20190401
torchvision-nightly==0.2.1
tqdm==4.31.1
traitlets==4.3.2
typing==3.6.6
typing-extensions==3.7.2
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.15.1
wrapt==1.11.1
